Fixed performance and logic of mount tower menu

+++ proj/src/Game/Gameplay/gameplay.c
@@ -64,7 +64,7 @@ static int8_t pause_current_selection = -1;
 static int8_t shop_current_selection = -1;
 static int8_t tower_current_selection = -1;
 static int8_t select_game_current_arena = -1;
-//static bool just_changed_selection = false;
+static bool just_changed_selection = false;
 
 bool multiplayer = false;
 
@@ -281,8 +281,8 @@ static void checkGameHovered(TowerArray* array) {
   if (pressed_game_button) {
     pressed_game_button = false;
     if(player1->hasWeapon){
-      float bullet_x = player1->x; 
-      float bullet_y = player1->y+player1->origin_offset_y - 10; 
+      float bullet_x = (float)player1->player->gameObject->x; 
+      float bullet_y = (float)player1->player->gameObject->y+player1->origin_offset_y; 
       int16_t damage = 200;
 
       Bullet* new_bullet = initializeBullet(bullet_x, bullet_y, mouse_x, mouse_y, damage);
@@ -689,6 +689,7 @@ static void checkTowerMenuHovered(ButtonArray* array) {
           }
         } else {
           selected_tower_index = 0;
+          just_changed_selection = true;
           Button* button = getButtonArray(&tower_buttons, selected_tower_index + 1);
           destroy_sprite(button->hovering);
           destroy_sprite(button->no_hovering);
@@ -718,6 +719,7 @@ static void checkTowerMenuHovered(ButtonArray* array) {
           }
         } else {
           selected_tower_index = 1;
+          just_changed_selection = true;
           Button* button = getButtonArray(&tower_buttons, selected_tower_index + 1);
           destroy_sprite(button->hovering);
           destroy_sprite(button->no_hovering);
@@ -730,9 +732,13 @@ static void checkTowerMenuHovered(ButtonArray* array) {
       case 3: 
         if (type_tower_menu) {
           // Upgrade turret damage
-          selected_tower_base->damage *= 1.1;
+          if (selected_tower_base->level < 5) {
+            selected_tower_base->damage *= 1.1;
+            selected_tower_base->level++;
+          }
         } else {
           selected_tower_index = 2;
+          just_changed_selection = true;
           Button* button = getButtonArray(&tower_buttons, selected_tower_index + 1);
           destroy_sprite(button->hovering);
           destroy_sprite(button->no_hovering);
@@ -790,7 +796,7 @@ static void checkTowerMenuHovered(ButtonArray* array) {
     }
   }
 
-  if (!type_tower_menu && selected_tower_index >= 0 && selected_tower_index <= 2) {
+  if (!type_tower_menu && selected_tower_index >= 0 && selected_tower_index <= 2 && just_changed_selection) {
     Button* button = getButtonArray(&tower_buttons, 4);
     destroy_sprite(button->hovering);
     destroy_sprite(button->no_hovering);
@@ -798,9 +804,10 @@ static void checkTowerMenuHovered(ButtonArray* array) {
     button->no_hovering = create_sprite((xpm_map_t)MountButton, 0, 0, false, true);      
   }
   
-  if (!type_tower_menu) {
+  if (!type_tower_menu && just_changed_selection) {
+    just_changed_selection = false;
     for (int i = 0; i < 2; i++) {
-      if (i != selected_tower_index) {
+      if (i != selected_tower_index && unlocked_turrets[i] == 1) {
         Button* button = getButtonArray(&tower_buttons, i+1);
         destroy_sprite(button->hovering);
         destroy_sprite(button->no_hovering);
@@ -1005,7 +1012,11 @@ static void updateTowerMenu() {
 
       pushButtonArray(&tower_buttons, initializeButton((xpm_map_t)LeftButton, (xpm_map_t) LeftButton, screen.xres/2 - 300, screen.yres/2 - 250, 0xFFFE, false, true));
       pushButtonArray(&tower_buttons, initializeButton((xpm_map_t)RightButton, (xpm_map_t)RightButton, screen.xres/2 - 300, screen.yres/2 - 100, 0xFFFE, false, true));
-      pushButtonArray(&tower_buttons, initializeButton((xpm_map_t)UpgradeButtonHovered, (xpm_map_t)UpgradeButton, screen.xres/2 - 300, screen.yres/2 + 50, 0xFFFE, false, true));
+      if (selected_tower_base->level < 5) {
+        pushButtonArray(&tower_buttons, initializeButton((xpm_map_t)UpgradeButtonHovered, (xpm_map_t)UpgradeButton, screen.xres/2 - 300, screen.yres/2 + 50, 0xFFFE, false, true));
+      } else {
+        pushButtonArray(&tower_buttons, initializeButton((xpm_map_t)UpgradeButtonUnclickable, (xpm_map_t)UpgradeButtonUnclickable, screen.xres/2 - 300, screen.yres/2 + 50, 0xFFFE, false, true));
+      }
       pushButtonArray(&tower_buttons, initializeButton((xpm_map_t)UnmountButtonHovered, (xpm_map_t)UnmountButton, screen.xres/2 - 300, screen.yres/2 + 200, 0xFFFE, false, true));
     }
